@@ -1,12 +1,12 @@
-package com.cybersoft.cozastore.entity;
+package com.cybersoft.cozaStore.entity;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+
 import java.util.Date;
 import java.util.List;
 
 @Entity(name = "size")
 public class SizeEntity {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
@@ -19,6 +19,7 @@ public class SizeEntity {
 
     @OneToMany(mappedBy = "size")
     private List<ProductEntity> products;
+
 
     public int getId() {
         return id;
