@@ -1,19 +1,16 @@
-package com.cybersoft.cozastore.security;
+package com.cybersoft.cozaStore.security;
 
-
-import com.cybersoft.cozastore.filter.JwtFilter;
-import com.cybersoft.cozastore.provider.CustomAuthenProvider;
+import com.cybersoft.cozaStore.filter.JwtFilter;
+import com.cybersoft.cozaStore.provider.CustomAuthenProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -23,52 +20,69 @@ import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
-@Configuration // Class sẽ được quét khi spring boot chạy ở tầng config
-@EnableWebSecurity // Custom Spring Secutiry
+@Configuration //Class sẽ được quét khi spring boot chạy ở tầng config
+@EnableWebSecurity // Custom Spring Security
+
+
+/**
+ * Tạo danh sách user ảo lưu trên thanh ram dùng để chứng thực
+ */
+
 public class SecurityConfig {
 
-    @Autowired
-    private JwtFilter jwtFilter;
-
-    // Tạo ra chuẩn mã hóa password là lưu trữ ở IOC
+    //TTạo ra chuẩn mã hoá password và lưu trên IOC
     @Bean
-    public PasswordEncoder passwordEncoder() {
-
+    public PasswordEncoder passwordEncoder(){
         return new BCryptPasswordEncoder();
     }
 
     @Autowired
+    private JwtFilter jwtFilter;
+    @Autowired
     CustomAuthenProvider customAuthenProvider;
 
-    // Chỉ định AuthenticationManager sử dụng CustomAuthenProvider
+    //chỉ định authenticationManager sử dụng CustomAuthenProvider
     @Bean
-    public AuthenticationManager authenticationManager(HttpSecurity httpSecurity) throws Exception {
-
+    public AuthenticationManager authenticationManager (HttpSecurity httpSecurity) throws Exception {
+        //CustomAuthenProvider customAuthenProvider = new CustomAuthenProvider();
         return httpSecurity.getSharedObject(AuthenticationManagerBuilder.class)
                 .authenticationProvider(customAuthenProvider)
                 .build();
     }
 
+//    @Bean
+//    public UserDetailsService userDetailsService(PasswordEncoder passwordEncoder){
+//        //Tạo tài khoản có tên là admin có role là admin
+//        UserDetails admin  = User.withUsername("admin").password(passwordEncoder.encode("123")).roles("ADMIN").build();
+//
+//        UserDetails user  = User.withUsername("user").password(passwordEncoder.encode("123")).roles("USER").build();
+//
+//        //Lưu trữ 2 user vừa tạo trên RAM
+//        return new InMemoryUserDetailsManager(admin, user);
+
+//    }
+
+    //Cấu hình chứng thực cho các link địa chỉ cho Securit
+    //GET/product: ai truy cập cũng được
+    //PUT/product: ADMIN mới truy cập được
+    //GET/cart: user mới truy cập được
+
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        return http.csrf().disable() // Tắt tấn công theo kiểu cross-site (gọi link theo nhiều trình duyệt khác nhau)
-                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS) // Không sử dụng session --- STATELESS: Không sử dụng cái gì hết
+        return http.csrf().disable() //Tắt tấn công theo kiểu cross-site(gọi link theo nhiều trình duyệt khác nhau)
+                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS) //không cho sử dụng session
                 .and()
                 .authorizeHttpRequests()
-
-                // các thằng dưới là con của thằng trên ----- Matchers là so sánh kiểm tra dữ liệu
-                .antMatchers("/login/**").permitAll() //permitALl() : nếu có .per thì link này ai gọi cũng được tất cả (ALL)
-                .antMatchers("/file/**").permitAll()
-                .antMatchers("/cart/**").permitAll()
-                .antMatchers(HttpMethod.POST, "/product").hasRole("ADMIN") // link /product với phương thức POST phải có role ADMIN mới truy cập được
-                .antMatchers(HttpMethod.GET, "/product").permitAll()
-                .antMatchers(HttpMethod.PUT, "/product").hasRole("ADMIN")
-                .antMatchers(HttpMethod.GET, "/cart").hasRole("USER")
-
-                .anyRequest().authenticated() // Tất cả các link còn lại cần phải chứng thực
+                        .requestMatchers("/login/**").permitAll() //permitAll: là khong cần chứng thực
+                        .requestMatchers("/file/**").permitAll()
+                        .requestMatchers(HttpMethod.POST, "/product").hasRole("ADMIN")// link /product với phương thúc post phải có role ADMIN mới truy cập được
+                        .requestMatchers(HttpMethod.GET, "/product").permitAll()
+                        .requestMatchers(HttpMethod.PUT, "/product").hasRole("ADMIN")
+                        .requestMatchers(HttpMethod.GET, "/cart").hasRole("USER")
+                        .anyRequest().authenticated() //Tất cả các link còn lại đều phải chứng thực
                 .and()
                 .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class)
                 .build();
+
     }
 }
-
