@@ -1,7 +1,7 @@
-package com.cybersoft.cozastore.controller;
+package com.cybersoft.cozaStore.controller;
 
-import com.cybersoft.cozastore.payload.BaseResponse;
-import com.cybersoft.cozastore.service.imp.FileServiceImp;
+import com.cybersoft.cozaStore.payload.response.Baseresponse;
+import com.cybersoft.cozaStore.service.imp.FileServiceImp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -15,32 +15,26 @@ import java.io.IOException;
 @RestController
 @RequestMapping("/file")
 public class FileController {
-
     @Autowired
     private FileServiceImp fileServiceImp;
 
     @PostMapping("")
     public ResponseEntity<?> uploadFile(@RequestParam MultipartFile file) throws IOException {
-
-        String message = fileServiceImp.uploadFile(file);
-
-        BaseResponse baseResponse = new BaseResponse();
-        baseResponse.setStatusCode(200);
-        baseResponse.setMessage("Thanh cong ");
-        baseResponse.setData(message);
-
-        return new ResponseEntity<>(baseResponse, HttpStatus.OK);
+       String message = fileServiceImp.uploadFile(file);
+        Baseresponse baseresponse = new Baseresponse();
+        baseresponse.setStatusCode(200);
+        baseresponse.setMessage("Thanh cong");
+        baseresponse.setData(message);
+        return new ResponseEntity<>(baseresponse, HttpStatus.OK);
     }
-
     @GetMapping("/{fileName}")
-    public ResponseEntity<?> downloadFile(@PathVariable String fileName) throws IOException {
-        byte[] fileImage = fileServiceImp.downloadFile(fileName);
+    public ResponseEntity<?> downloadFile(@PathVariable String fileName){
+        byte[] fileImage =  fileServiceImp.downloadFile(fileName);
 
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.IMAGE_PNG);
 
 
-        return new ResponseEntity<>(fileImage, headers, HttpStatus.OK);
-
+        return new ResponseEntity<> (fileImage, headers, HttpStatus.OK);
     }
 }
