@@ -1,14 +1,11 @@
-package com.cybersoft.cozastore.controller;
+package com.cybersoft.cozaStore.controller;
 
-import com.cybersoft.cozastore.payload.BaseResponse;
-import com.cybersoft.cozastore.payload.request.SignUpRequest;
-import com.cybersoft.cozastore.service.imp.LoginServiceImp;
-import com.cybersoft.cozastore.util.JwtHelper;
+import com.cybersoft.cozaStore.payload.response.Baseresponse;
+import com.cybersoft.cozaStore.payload.request.SignUpRequest;
+import com.cybersoft.cozaStore.service.imp.LoginServiceImp;
+import com.cybersoft.cozaStore.util.JwtHelper;
 import com.google.gson.Gson;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.io.Encoders;
-import io.jsonwebtoken.security.Keys;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,11 +16,10 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
-import javax.crypto.SecretKey;
-import javax.validation.Valid;
 import java.util.List;
 
 @CrossOrigin
+
 @RestController
 @RequestMapping("/login")
 public class LoginController {
@@ -41,39 +37,42 @@ public class LoginController {
 
     @PostMapping("/signin")
     public ResponseEntity<?> signin(@RequestParam String email, @RequestParam String password){
-//        SecretKey key = Keys.secretKeyFor(SignatureAlgorithm.HS256);
-//        String secretString = Encoders.BASE64.encode(key.getEncoded());
-//        System.out.printf("kiemtra " + secretString); // kiểm tra key
+        /*SecretKey key =  Keys.secretKeyFor(SignatureAlgorithm.HS256);
+        String secretString = Encoders.BASE64.encode(key.getEncoded());*/
 
         UsernamePasswordAuthenticationToken authen = new UsernamePasswordAuthenticationToken(email, password);
+
+        //System.out.println("Kiem tra " + secretString);
+
         authenticationManager.authenticate(authen);
 
-        // Lấy danh sách role dã lưu từ Security Context Holder khi AuthenManager chứng thực thành công
+        //lấy danh sách role đã lưu trữ từ security context holder khi authemanager chứng thực thành công
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         List<SimpleGrantedAuthority> roles = (List<SimpleGrantedAuthority>) authentication.getAuthorities();
-
         String jsonRole = gson.toJson(roles);
-
+//        for (SimpleGrantedAuthority data: roles) {
+//            System.out.printf("role"  + data);
+//        }
         String token = jwtHelper.generateToken(jsonRole);
 
-        BaseResponse baseResponse = new BaseResponse();
-        baseResponse.setStatusCode(200);
-        baseResponse.setMessage("");
-        baseResponse.setData(token);
+        Baseresponse baseresponse = new Baseresponse();
+        baseresponse.setStatusCode(200);
+        baseresponse.setMessage("");
+        baseresponse.setData(token);
 
-        return new ResponseEntity<>(baseResponse, HttpStatus.OK);
+
+        return new ResponseEntity<>(baseresponse, HttpStatus.OK);
     }
 
     @PostMapping("/signup")
     public ResponseEntity<?> signup(@Valid @RequestBody SignUpRequest signUpRequest){
         boolean isSuccess = loginServiceImp.insertUser(signUpRequest);
 
-        BaseResponse baseResponse = new BaseResponse();
-        baseResponse.setStatusCode(200);
-        baseResponse.setMessage("");
-        baseResponse.setData(isSuccess);
+        Baseresponse baseresponse = new Baseresponse();
+        baseresponse.setStatusCode(200);
+        baseresponse.setMessage("");
+        baseresponse.setData(isSuccess);
 
-        return new ResponseEntity<>(baseResponse, HttpStatus.OK);
+        return new ResponseEntity<> (baseresponse, HttpStatus.OK);
     }
-
 }
