@@ -1,9 +1,9 @@
-package com.cybersoft.cozastore.controller;
+package com.cybersoft.cozaStore.controller;
 
-import com.cybersoft.cozastore.service.ProductService;
-import com.cybersoft.cozastore.service.imp.ProductServiceImp;
+import com.cybersoft.cozaStore.service.imp.ProductSeviceImp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -15,27 +15,22 @@ import java.io.IOException;
 public class ProductController {
 
     @Autowired
-    private ProductServiceImp productServiceImp;
+    private ProductSeviceImp productSeviceImp;
 
     @GetMapping("")
     public ResponseEntity<?> getAllProduct(){
 
-        return new ResponseEntity<>("Hello product", HttpStatus.OK);    }
-
-    @PostMapping("")
-    public ResponseEntity<?> insertProdut(@RequestParam String name, @RequestParam MultipartFile file,
-                                          @RequestParam double price, @RequestParam int quantity, @RequestParam int idColor,
-                                          @RequestParam int idSize, @RequestParam int idCategory
-    ) throws IOException {
-
-        boolean isSuccess = productServiceImp.insertProduct(name, file, price, quantity, idColor, idSize, idCategory);
-
-        return new ResponseEntity<>("Insert Product", HttpStatus.OK);
+        return new ResponseEntity<>("Hello product", HttpStatus.OK);
     }
 
-    @PutMapping("")
-    public ResponseEntity<?> addProduct(){
+    @PostMapping("")
+    public ResponseEntity<?> insertProduct(@RequestParam String name,
+                                           @RequestParam MultipartFile file, @RequestParam double price,
+                                           @RequestParam int quantity, @RequestParam int idColor,
+                                           @RequestParam int idSize, @RequestParam int idCategory) throws IOException {
 
-        return new ResponseEntity<>("Hello add", HttpStatus.OK);
+        boolean isSussecc = productSeviceImp.insertProduct(name, file, price, quantity, idColor, idSize, idCategory);
+
+        return new ResponseEntity<>("Insert Product", HttpStatus.OK);
     }
 }
