@@ -1,9 +1,12 @@
-package com.cybersoft.cozastore.filter;
+package com.cybersoft.cozaStore.filter;
 
-
-import com.cybersoft.cozastore.util.JwtHelper;
+import com.cybersoft.cozaStore.util.JwtHelper;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
@@ -13,19 +16,13 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 
-// dành cho việc giải mã token
-// Tạo filter để hứng token mỗi khi người dùng gọi request
-
-@Service //combonent cũng được
+@Service
+//Tạo filter để hứng token mỗi khi nguời dùng gọi request
 public class JwtFilter extends OncePerRequestFilter {
 
     @Autowired
@@ -35,36 +32,30 @@ public class JwtFilter extends OncePerRequestFilter {
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-
-        // Lấy token mà client truyền trên header (authorization)
-        String headerValue = request.getHeader("Authorization");
+        //lấy token mà client truyền trên header (Authorization)
+        String headerValue =request.getHeader("Authorization");
         if(headerValue != null && headerValue.startsWith("Bearer ")){
-
-            // Cắt chữ bearer để lấy token
-            String token = headerValue.substring(7);
-            jwtHelper.parserToken(token);
-            String data = jwtHelper.parserToken(token);
-            System.out.println("Kiem tra " + data);
-            if(data != null && !data.isEmpty()){
-                // Chứng thực hợp lệ, tạo chứng thực cho Security
-
-                Type listType = new TypeToken<ArrayList<SimpleGrantedAuthority>>(){}.getType();
-                List<SimpleGrantedAuthority> roles = gson.fromJson(data,listType);
+            //cắt chữ Bearer để lấy token
+           String token = headerValue.substring(7);
+           String data = jwtHelper.parseToken(token);
+            System.out.printf("Kiem tra " + data);
+            if(data != null && data.isEmpty()){
+                Type listType = new TypeToken<ArrayList< SimpleGrantedAuthority>>(){}.getType();
+                List<SimpleGrantedAuthority> roles = gson.fromJson(data, listType);
 
 //                List<GrantedAuthority> roles = new ArrayList<>();
 //                GrantedAuthority grantedAuthority = new SimpleGrantedAuthority("ROLE_ADMIN");
+//                roles.add( grantedAuthority);
 
-//                roles.add(roles);
 
-                UsernamePasswordAuthenticationToken user =
-                        new UsernamePasswordAuthenticationToken("","",roles);
-                SecurityContext context = SecurityContextHolder.getContext();
+               // Chứng thực hợp lệ và tạo chứng thực cho security
+                UsernamePasswordAuthenticationToken user = new UsernamePasswordAuthenticationToken("", "", roles);
+                SecurityContext context = SecurityContextHolder .getContext();
                 context.setAuthentication(user);
-
             }
-        } else {
-            // không hợp lệ
-            System.out.println("Nội dung header không hợp lệ");
+        }else{
+            //khong hop le
+            System.out.println("Noi dung khong hop le");
         }
 
         filterChain.doFilter(request, response);
