@@ -1,12 +1,14 @@
-package com.cybersoft.cozastore.repository;
+package com.cybersoft.cozaStore.repository;
 
-import com.cybersoft.cozastore.entity.UserEntity;
+import com.cybersoft.cozaStore.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 
-@Repository // đưa lên ioc (container dùng chung)
-public interface UserRepository extends JpaRepository<UserEntity,Integer> {
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
     UserEntity findByEmail(String email);
-
+    List<UserEntity> findByUsernameAndPassword(String username, String password);
 }
