@@ -1,4 +1,4 @@
-package com.cybersoft.cozastore.service.imp;
+package com.cybersoft.cozaStore.service.imp;
 
 import org.springframework.web.multipart.MultipartFile;
 
@@ -7,9 +7,6 @@ import java.io.IOException;
 public interface FileServiceImp {
 
     String uploadFile(MultipartFile file) throws IOException;
-    // upload file lên sẽ là bỏ file hoặc không bỏ file nếu nó không chạy nữa thì nó throws exeception
 
-    byte[] downloadFile(String fileName) throws IOException;
-
-
+    byte[] downloadFile(String fileName) ;
 }
